Youtube_Notes_Generator/requirements.txt
google-generativeai
streamlit
youtube_transcript_api
dotenv